Extract useSpeedKatakanaGame hook and FlashFeedback from Game
The game loop (word cursor, kana input commit path, correct/skip
accounting, flash reveal, session stats) moves into
use-speed-katakana-game.ts; the duplicated five-line flash block becomes
one showFlash(skipped) with a named FLASH_DURATION_MS. Game.tsx keeps the
data fetch and rendering only.
