+++ src/components/screens/SpeedKatakanaScreen/Game.tsx
@@ -1,29 +1,15 @@
-import { useEffect, useMemo, useRef, useState } from "react";
-import { translateValue } from "@/lib/wanakana-adapter";
-import { isMatch, isOnTrack } from "@/lib/speed-katakana-match";
+import { useMemo } from "react";
 import { Input } from "@/components/ui/input";
 import { DefaultErrorFallback } from "@/components/error";
 import KaomojiAnimation from "@/components/common/KaomojiLoading";
 import { useJsonFetch } from "@/hooks/use-json";
-import { useSpeak } from "@/hooks/use-jp-speak";
-import { useCorrectSound } from "@/hooks/use-correct-sound";
-import { useKanaInput } from "@/hooks/use-kana-input";
 import assetsPaths from "@/lib/assets-paths";
-import {
-  isForgotCommand,
-  isForgotCommandPrefix,
-} from "@/lib/practice-commands";
 import { EndSession } from "@/components/shared-practice/EndSessionButton";
 import { ChallengeSetData, SessionStats, SpeedKatakanaSettings } from "./types";
 import { randomFontIndex } from "@/hooks/use-change-font";
-import { percent, shuffle } from "@/lib/utils";
-
-type GameWord = {
-  katakana: string;
-  english: string;
-  /** Index into the --jap-font-* CSS vars, or null to use the default font. */
-  fontIndex: number | null;
-};
+import { shuffle } from "@/lib/utils";
+import { GameWord, useSpeedKatakanaGame } from "./use-speed-katakana-game";
+import { FlashFeedback } from "./FlashFeedback";
 
 export const Game = ({
   settings,
@@ -55,44 +41,12 @@ export const Game = ({
     settings.wordCount,
   ]);
 
-  const [index, setIndex] = useState(0);
-  const [inputValue, setInputValue] = useState("");
-  const [flash, setFlash] = useState<{
-    english: string;
-    key: number;
-    skipped: boolean;
-  } | null>(null);
-  const inputRef = useRef<HTMLInputElement | null>(null);
-
-  // Stats live in refs so per-keystroke bookkeeping doesn't trigger re-renders.
-  const startTimeRef = useRef<number | null>(null);
-  const correctCharsRef = useRef(0);
-  const errorsRef = useRef(0);
-  const inErrorStateRef = useRef(false);
-  const flashKeyRef = useRef(0);
-  const flashTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-
-  // Sound feedback. useSpeak is a hook, so it must run before any early return;
-  // an empty word is harmless while the set is still loading.
-  const speak = useSpeak(words[index]?.katakana ?? "");
-  const playCorrect = useCorrectSound();
-  const kanaInput = useKanaInput({
-    setRawValue: setInputValue,
-    onCommit: (raw) => handleChange(raw),
+  const game = useSpeedKatakanaGame({
+    words,
+    settings,
+    onProgress,
+    onComplete,
   });
-
-  // Keep the input focused as words advance and once the set has loaded.
-  useEffect(() => {
-    inputRef.current?.focus();
-  }, [index, words.length]);
-
-  // Effect needed: clears any pending flash timeout on unmount.
-  useEffect(() => {
-    const timer = flashTimerRef;
-    return () => {
-      if (timer.current) clearTimeout(timer.current);
-    };
-  }, []);
 
   if (status === "error") {
     return (
@@ -112,117 +66,7 @@ export const Game = ({
     );
   }
 
-  const current = words[index];
-
-  const playFeedback = () => {
-    playCorrect({
-      enabled: settings.sound.enabled,
-      speak:
-        settings.sound.enabled && settings.sound.type === "speak"
-          ? speak
-          : undefined,
-    });
-  };
-
-  const advance = () => {
-    const nextIndex = index + 1;
-    if (nextIndex >= words.length) {
-      finish();
-      return;
-    }
-    setIndex(nextIndex);
-    setInputValue("");
-    onProgress((nextIndex / words.length) * 100);
-  };
-
-  const handleSkip = () => {
-    // Skipping counts against accuracy so it isn't a free pass.
-    errorsRef.current += 1;
-    inErrorStateRef.current = false;
-    if (!settings.displayEnglish) {
-      if (flashTimerRef.current) clearTimeout(flashTimerRef.current);
-      flashKeyRef.current += 1;
-      setFlash({
-        english: current.english,
-        key: flashKeyRef.current,
-        skipped: true,
-      });
-      flashTimerRef.current = setTimeout(() => setFlash(null), 1200);
-    }
-    advance();
-    inputRef.current?.focus();
-  };
-
-  const handleChange = (raw: string) => {
-    if (startTimeRef.current === null) startTimeRef.current = Date.now();
-
-    // Keep "skip" / "forgot" as latin so the keyboard command stays typable.
-    if (isForgotCommandPrefix(raw) || isForgotCommand(raw)) {
-      setInputValue(raw.replace(/\s+/g, ""));
-      return;
-    }
-
-    const converted = translateValue(raw, "katakana");
-    const target = current.katakana;
-
-    // Matching happens in romaji space (see lib/speed-katakana-match) so the
-    // long-vowel mark, IME whitespace, and mid-syllable typing behave.
-    const onTrack = isOnTrack(converted, target);
-    if (!onTrack && !inErrorStateRef.current) {
-      errorsRef.current += 1;
-    }
-    inErrorStateRef.current = !onTrack;
-
-    if (isMatch(converted, target)) {
-      correctCharsRef.current += target.length;
-      inErrorStateRef.current = false;
-      playFeedback();
-      if (!settings.displayEnglish) {
-        if (flashTimerRef.current) clearTimeout(flashTimerRef.current);
-        flashKeyRef.current += 1;
-        setFlash({
-          english: current.english,
-          key: flashKeyRef.current,
-          skipped: false,
-        });
-        flashTimerRef.current = setTimeout(() => setFlash(null), 1200);
-      }
-      advance();
-      return;
-    }
-
-    setInputValue(converted);
-  };
-
-  const handleKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
-    if (e.nativeEvent.isComposing) return;
-
-    if (e.key === "Escape") {
-      e.preventDefault();
-      handleSkip();
-      return;
-    }
-
-    if (e.key === "Enter" && isForgotCommand(inputValue)) {
-      e.preventDefault();
-      handleSkip();
-    }
-  };
-
-  const finish = () => {
-    const elapsedMs = startTimeRef.current
-      ? Date.now() - startTimeRef.current
-      : 0;
-    const minutes = elapsedMs / 60000;
-    const charsPerMinute =
-      minutes > 0 ? Math.round(correctCharsRef.current / minutes) : 0;
-
-    const attempts = correctCharsRef.current + errorsRef.current;
-    const accuracy = percent(correctCharsRef.current, attempts, 100);
-
-    onProgress(100);
-    onComplete({ accuracy, charsPerMinute });
-  };
+  const current = game.current;
 
   return (
     <div className="flex flex-col w-full h-full gap-4 mx-auto [@media(min-height:900px)]:justify-center animate-fade-in-fast">
@@ -231,12 +75,12 @@ export const Game = ({
       <div className="flex flex-col items-center justify-center flex-1 min-h-0 [@media(min-height:900px)]:flex-none gap-3 text-center">
         <div className="flex items-center justify-center gap-1 pt-4">
           <span className="px-2 text-xs font-bold tabular-nums">
-            {index + 1} / {words.length}
+            {game.index + 1} / {game.wordCount}
           </span>
           <button
             className="text-xs font-bold tracking-wide text-red-500 transition-opacity -translate-y-[0.5px] hover:opacity-70"
             tabIndex={-1}
-            onClick={handleSkip}
+            onClick={game.handleSkip}
           >
             {`>>`}
           </button>
@@ -261,23 +105,14 @@ export const Game = ({
           </p>
         )}
         <div className="flex flex-col items-center gap-1 pb-4">
-          <div className="h-4">
-            {!settings.displayEnglish && flash && (
-              <span
-                key={flash.key}
-                className={`text-xs font-bold tracking-wide uppercase animate-english-flash ${flash.skipped ? "text-muted-foreground" : "text-green-500"}`}
-              >
-                {flash.skipped ? "→" : "✓"} {flash.english}
-              </span>
-            )}
-          </div>
+          <FlashFeedback flash={settings.displayEnglish ? null : game.flash} />
         </div>
       </div>
 
       <div className="shrink-0 pl-3 pr-1 pb-[max(0.5rem,env(safe-area-inset-bottom))]">
         <Input
-          ref={inputRef}
-          value={inputValue}
+          ref={game.inputRef}
+          value={game.inputValue}
           autoFocus
           autoCapitalize="off"
           autoCorrect="off"
@@ -286,8 +121,8 @@ export const Game = ({
           aria-label='Type romaji or type "skip"'
           placeholder='Type romaji or "skip"'
           className="w-full text-2xl text-center border-2 z-1000 rounded-2xl h-14"
-          onKeyDown={handleKeyDown}
-          {...kanaInput}
+          onKeyDown={game.handleKeyDown}
+          {...game.kanaInput}
         />
       </div>
     </div>
